requirements.txt
authlib ~= 1.2
django ~= 4.2
python-dotenv ~= 1.0
requests ~= 2.31
djangorestframework==3.15.2
python-dateutil==2.9.0
Pillow==10.4.0
djangorestframework-simplejwt==5.3.1 
pyjwt==2.9.0
numpy==2.1.1 
opencv-python==4.10.0.84
protobuf==3.12.4
psycopg2-binary==2.9.9
whitenoise==6.7.0
gunicorn==23.0.0
dj-database-url==2.2.0
black==24.8.0
asgiref==3.8.1
certifi==2024.8.30
django-cors-headers==4.4.0